handle shared view
